requirements.txt
altgraph==0.17
appdirs==1.4.3
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.5
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cmake==3.17.2
colorama==0.4.3
cryptography==2.3.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
entrypoints==0.3
filelock==3.0.12
future==0.18.2
h5py==2.10.0
idna==2.7
importlib-metadata==1.6.0
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
Jinja2==2.11.2
jsonschema==2.6.0
jupyter-client==6.1.3
jupyter-core==4.6.3
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
more-itertools==8.2.0
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.3
olefile==0.46
open3d-python==0.7.0.0
opencv-contrib-python==3.4.2.16
opencv-python==3.4.2.16
packaging==20.3
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pefile==2019.4.18
pickleshare==0.7.5
Pillow==5.2.0
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.5
psutil==5.4.7
py==1.8.1
pycparser==2.20
Pygments==2.6.1
PyInstaller==3.6
pyOpenSSL==18.0.0
pyparsing==2.4.7
PySocks==1.6.8
pytest==5.4.1
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==3.13
pyzmq==19.0.1
QtPy==1.9.0
requests==2.19.1
Send2Trash==1.5.0
six==1.11.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.2
urllib3==1.23
virtualenv==20.0.18
wcwidth==0.1.9
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.1.0
